requirements.txt
langchain-community
langchain_openai
streamlit
python-dotenv
pypdf2
streamlit_extras
openai
tiktoken
transformers